changing web-server to use debian instead of alpine + optimized puppeteer usage
I couldn't get puppeteer working properly on the node:22-alpine image. Moving to node:22-bullseye seems to have fixed most of the issues. I also removed creating an account for sandboxing, as we don't need sandboxing in this app (this also makes building the image take a lot less time).

Also changed it so we aren't instantiating a new instance of chrome every time we receive a request to render text (we now have a single chrome instance and open a new tab for each job)